backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
pydantic==2.5.3
PyYAML==6.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1 